+++ src/views/source.usage.js
@@ -3,7 +3,7 @@ const codeExample = 'let name = "world";\n\nconsole.log(`Hello, ${name}!`);';
 export default {
     demo: {
         view: 'source',
-        data: { content: codeExample, syntax: 'js' }
+        data: { syntax: 'js', content: codeExample }
     },
     examples: [
         {
@@ -21,6 +21,23 @@ export default {
                 'md:"More syntaxes may be added via `import \'codemirror/mode/[name]/[name]\';`"'
             ],
             source: false
+        },
+        {
+            title: 'Highlight ranges',
+            demo: {
+                view: 'source',
+                data: {
+                    syntax: 'js',
+                    content: codeExample,
+                    refs: [
+                        { range: [4, 8] },
+                        { range: [21, 28], type: 'link', href: '#example', tooltip: {
+                            position: 'trigger',
+                            content: ['text:"Link to "', 'text:href']
+                        } }
+                    ]
+                }
+            }
         },
         {
             title: 'Custom line numbers',
@@ -41,6 +58,18 @@ export default {
                     content: codeExample,
                     syntax: 'js',
                     lineNum: false
+                }
+            }
+        },
+        {
+            title: 'Max content size for highlight',
+            beforeDemo: ['md:"By default a syntax highlighing is not appling to a source which is bigger than 250Kb. Option `maxSourceSizeToHighlight` is using to change max size of source to be highlighted."'],
+            demo: {
+                view: 'source',
+                data: {
+                    content: codeExample,
+                    syntax: 'js',
+                    maxSourceSizeToHighlight: 4
                 }
             }
         }